requirements.txt
numpy
scipy
matplotlib
Pillow
quickdraw